Move module C header files to a modules subfolder
