+++ Support/Sources/COM_ABI/include/SWRT/modules/COM.h
